+++ app/helpers/application_helper.rb
@@ -1,4 +1,7 @@
 # frozen_string_literal: true
 
 module ApplicationHelper
+  def l(object, options = {})
+    super(object, options) if object
+  end
 end