+++ app/helpers/codelist_helper.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module CodelistHelper
+  def yaml_to_options(entity, type)
+    yaml = load_yaml(entity, type)
+    data = yaml["data"]
+    data.collect { |item| [item["name"], item["code"]] }
+  end
+
+  def load_yaml(entity, type)
+    YAML.safe_load(File.read("#{Rails.root}/vendor/data/codelists/IATI/#{IATI_VERSION}/#{entity}/#{type}.yml"))
+  end
+end