+++ db/migrate/20191118095803_create_funds.rb
@@ -1,0 +1,9 @@
+class CreateFunds < ActiveRecord::Migration[6.0]
+  def change
+    create_table :funds, id: :uuid do |t|
+      t.string :name
+      t.references :organisation, type: :uuid
+      t.timestamps
+    end
+  end
+end