+++ spec/models/activity_spec.rb
@@ -1,0 +1,11 @@
+require "rails_helper"
+
+RSpec.describe Activity, type: :model do
+  describe "relations" do
+    it { should belong_to(:hierarchy) }
+  end
+
+  describe "constraints" do
+    it { should validate_uniqueness_of(:identifier) }
+  end
+end